Refactor group key and name types into separate settingsStoreTypes file
