+++ src/store/settingsStoreTypes.d.ts
@@ -1,0 +1,22 @@
+export type GroupKey = 'dean' | 'lab' | 'proj' | 'comp';
+
+export type GroupName =
+  | 'Dziekańska'
+  | 'Komputerowa'
+  | 'Laboratoryjna'
+  | 'Projektowa';
+
+export interface SettingsState {
+  groups: Record<GroupKey, string | undefined>;
+  options: Record<GroupKey, string[]>;
+  loading: boolean;
+  lastFetchedDean: string | null;
+  activeDropdown: GroupKey | null;
+
+  actions: {
+    setGroup: (key: GroupKey, value: string) => void;
+    fetchInitialDeanGroups: () => Promise<void>;
+    fetchDependentGroups: (deanGroup: string) => Promise<void>;
+    setActiveDropdown: (key: GroupKey | null) => void;
+  };
+}
